refactor: move to shared::Postgres
