+++ service/src/lib.rs
@@ -180,6 +180,9 @@ pub use error::Error;
 
 pub mod logger;
 
+#[cfg(feature = "sqlx-postgres")]
+pub mod shared;
+
 #[cfg(feature = "secrets")]
 pub mod secrets;
 #[cfg(feature = "secrets")]
@@ -275,45 +278,6 @@ pub trait Factory: Send + Sync {
 pub trait ResourceBuilder<T> {
     fn new() -> Self;
     async fn build(self, factory: &mut dyn Factory, runtime: &Runtime) -> Result<T, crate::Error>;
-}
-
-#[cfg(feature = "sqlx-postgres")]
-pub struct Postgres;
-
-/// Get an `sqlx::PgPool` from any factory
-#[cfg(feature = "sqlx-postgres")]
-#[async_trait]
-impl ResourceBuilder<sqlx::PgPool> for Postgres {
-    fn new() -> Self {
-        Self {}
-    }
-
-    async fn build(
-        self,
-        factory: &mut dyn Factory,
-        runtime: &Runtime,
-    ) -> Result<sqlx::PgPool, crate::Error> {
-        use error::CustomError;
-
-        let connection_string = factory
-            .get_sql_connection_string(database::Type::Shared)
-            .await?;
-
-        // A sqlx Pool cannot cross runtime boundaries, so make sure to create the Pool on the service end
-        let pool = runtime
-            .spawn(async move {
-                sqlx::postgres::PgPoolOptions::new()
-                    .min_connections(1)
-                    .max_connections(5)
-                    .connect(&connection_string)
-                    .await
-            })
-            .await
-            .map_err(CustomError::new)?
-            .map_err(CustomError::new)?;
-
-        Ok(pool)
-    }
 }
 
 /// A tokio handle the service was started on
